requirements.txt
streamlit==1.42.2
google-genai==1.2.0
python-dotenv==1.0.1
google-generativeai==0.8.4